coupon and edit product

+++ NykantAPI/Models/Coupon.cs
@@ -10,7 +10,7 @@ namespace NykantAPI.Models
     {
         [Key]
         public string Code { get; set; }
-        public double Discount { get; set; }
+        public int Discount { get; set; }
         public bool Enabled { get; set; }
         public bool ForAllProducts { get; set; }
         public IEnumerable<CouponForProduct> CouponForProducts { get; set; }
